Remove transaction ID from history list
- Seperate row component out into its own file
- Remove unnecessary locale string
