+++ frontend/public/locales/en/common.json
@@ -138,7 +138,6 @@
     "checkout": "Checkout",
     "confirm-selection": "Confirm Selection",
     "created": "Created",
-    "id": "ID",
     "loading-saved-payment-methods": "Loading saved payment methods…",
     "payment": "Payment",
     "make-donation": "Make Donation",
