+++ frontend/src/components/payment/transactions/TransactionList.module.scss
@@ -1,5 +1,5 @@
 .transactionList {
-  --columns: 7;
+  --columns: 6;
 
   display: grid;
   grid-template-columns: repeat(1fr, var(--columns));
@@ -13,25 +13,5 @@
     grid-column: 1 / calc(var(--columns) + 1);
 
     align-items: center;
-  }
-
-  .info {
-    background-color: var(--bg-color-secondary);
-    border-radius: var(--border-radius);
-
-    padding: 10px;
-    margin: 5px 0;
-
-    min-width: 200px;
-
-    box-shadow: 0 0 0 1px rgba(0, 0, 0, 0.03), 0 1px 3px 1px rgba(0, 0, 0, 0.07),
-      0 2px 6px 2px rgba(0, 0, 0, 0.03);
-
-    .actions {
-      display: flex;
-      flex-direction: column;
-      align-items: flex-start;
-      gap: 10px;
-    }
   }
 }
