+++ frontend/src/components/payment/transactions/TransactionList.tsx
@@ -1,10 +1,8 @@
 import { useTranslation } from 'next-i18next'
-import Link from 'next/link'
-import { FunctionComponent, useState } from 'react'
+import { FunctionComponent } from 'react'
 import { Transaction } from '../../../types/Payment'
-import Button from '../../Button'
-import TransactionCancelButton from './TransactionCancelButton'
 import styles from './TransactionList.module.scss'
+import TransactionListRow from './TransactionListRow'
 
 interface ListProps {
   transactions: Transaction[]
@@ -19,7 +17,6 @@ const TransactionList: FunctionComponent<ListProps> = ({ transactions }) => {
         <tr>
           <td>{t('created')}</td>
           <td>{t('updated')}</td>
-          <td>{t('id')}</td>
           <td>{t('type')}</td>
           <td>{t('value')}</td>
           <td>{t('status')}</td>
@@ -29,7 +26,10 @@ const TransactionList: FunctionComponent<ListProps> = ({ transactions }) => {
       <tbody>
         {transactions.length ? (
           transactions.map((transaction) => (
-            <TransactionInfo key={transaction.id} transaction={transaction} />
+            <TransactionListRow
+              key={transaction.id}
+              transaction={transaction}
+            />
           ))
         ) : (
           <tr>
@@ -38,61 +38,6 @@ const TransactionList: FunctionComponent<ListProps> = ({ transactions }) => {
         )}
       </tbody>
     </table>
-  )
-}
-
-interface RowProps {
-  transaction: Transaction
-}
-
-const TransactionInfo: FunctionComponent<RowProps> = ({ transaction }) => {
-  const { t, i18n } = useTranslation()
-
-  const { id, created, updated, kind, value, status } = transaction
-
-  // Status may change through interaction
-  const [shownStatus, setStatus] = useState(status)
-
-  const prettyCreated = new Date(created * 1000).toLocaleString()
-  const prettyUpdated = new Date(updated * 1000).toLocaleString()
-  const prettyValue = new Intl.NumberFormat(i18n.language, {
-    style: 'currency',
-    currency: 'USD',
-    currencyDisplay: 'symbol',
-  }).format(value / 100)
-
-  const needsAttention = ['new', 'retry'].includes(shownStatus)
-
-  // Date object expects milliseconds since epoch
-  return (
-    <tr className={styles.info}>
-      <td>{prettyCreated}</td>
-      <td>{prettyUpdated}</td>
-      <td>{id}</td>
-      <td>{kind}</td>
-      <td>{prettyValue}</td>
-      <td>{status}</td>
-      <td className={styles.actions}>
-        <Link
-          href={`/payment/${
-            needsAttention ? transaction.id : `details/${transaction.id}`
-          }`}
-          passHref
-        >
-          <Button type={needsAttention ? 'primary' : 'secondary'}>
-            {needsAttention ? t('checkout') : t('view')}
-          </Button>
-        </Link>
-        {needsAttention ? (
-          <TransactionCancelButton
-            id={transaction.id}
-            onSuccess={() => setStatus('cancelled')}
-          />
-        ) : (
-          <></>
-        )}
-      </td>
-    </tr>
   )
 }
