+++ frontend/src/components/payment/transactions/TransactionListRow.module.scss
@@ -1,0 +1,19 @@
+.row {
+  background-color: var(--bg-color-secondary);
+  border-radius: var(--border-radius);
+
+  padding: 8px;
+  margin: 8px 0;
+
+  min-width: 200px;
+
+  box-shadow: 0 0 0 1px rgba(0, 0, 0, 0.03), 0 1px 3px 1px rgba(0, 0, 0, 0.07),
+    0 2px 6px 2px rgba(0, 0, 0, 0.03);
+
+  .actions {
+    display: flex;
+    flex-direction: column;
+    align-items: flex-start;
+    gap: 8px;
+  }
+}
